Add canDeleteMessage() method in MessageService : cannot delete the first message if there are any answers

+++ src/Controller/MessageController.php
@@ -104,6 +104,12 @@ class MessageController extends BaseController
         $thread = $message->getThread();
         $forum = $thread->getForum();
 
+        if (!$messageService->canDeleteMessage($message)) {
+            return $this->redirectToRoute('thread.show', [
+                'slug' => $thread->getSlug()
+            ]);
+        }
+
         $lastMessage = $messageService->deleteMessage($message);
 
         if (!$lastMessage) {
